+++ common-service/src/main/java/com/le07/commonservice/app/model/Biz.java
@@ -13,7 +13,7 @@ import java.util.Map;
 /**
  * 简述
  * <p/>
- * Created with IntelliJ IDEA.
+ * Created with IDEA
  * User: jh
  * Date: 13-6-28
  * Time: 上午10:29
@@ -25,7 +25,7 @@ public class Biz implements Statusable, Serializable {
     private static final long serialVersionUID = -6236818150499308790L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column(name = "biz_key", length = 32, nullable = false, unique = true)
     private String key;
@@ -47,11 +47,11 @@ public class Biz implements Statusable, Serializable {
     @Column(nullable = false)
     private Status status = Status.ENABLED;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
